Update camera zoom and character movement mechanics
This commit includes several tweaks to in-game camera and character's physics properties. The camera zoom level has now been halved to provide a wider field of view, and the level of friction has been reduced from 1 to 0.4 to improve the fluidity of character's movements. Also, the character has some changes in its forces to handle different states including jumping or when it's moving right or left.

The camera now follows the player character with a smoother movement as I added linear interpolation between the camera's current location and target location (player's location). These changes are intended to create a more engaging player experience.

Moreover, it disables the shake function of the camera for now. Lastly, unnecessary import statements are removed to improve code readability and performance.